// src/ontologies/rdf.d.ts
import { NamedNode } from '../models.js';
export declare var _ontologyResource: NamedNode;
export declare const rdf: {
    _ontologyResource: NamedNode;
    langString: NamedNode;
    type: NamedNode;
    Property: NamedNode;
    List: NamedNode;
    rest: NamedNode;
    first: NamedNode;
    nil: NamedNode;
};